+++ lib/src/inputs/src/pickers/date/range.dart
@@ -116,7 +116,7 @@ class ThemedDateRangePicker extends StatefulWidget {
 }
 
 class _ThemedDateRangePickerState extends State<ThemedDateRangePicker> {
-  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.of(context);
+  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.maybeOf(context);
   bool get isDark => Theme.of(context).brightness == Brightness.dark;
 
   String? get _parsedName {
@@ -217,7 +217,7 @@ class _ThemedDateRangePickerState extends State<ThemedDateRangePicker> {
   }
 
   String t(String key, [Map<String, dynamic> args = const {}]) {
-    String result = LayrzAppLocalizations.of(context)?.t(key, args) ?? widget.translations[key] ?? key;
+    String result = LayrzAppLocalizations.maybeOf(context)?.t(key, args) ?? widget.translations[key] ?? key;
 
     if (widget.overridesLayrzTranslations) {
       result = widget.translations[key] ?? key;