+++ lib/src/inputs/src/pickers/time/single.dart
@@ -124,7 +124,7 @@ class ThemedTimePicker extends StatefulWidget {
 }
 
 class _ThemedTimePickerState extends State<ThemedTimePicker> {
-  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.of(context);
+  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.maybeOf(context);
   String get pattern => widget.pattern ?? (widget.use24HourFormat ? '%H:%M' : '%I:%M %p');
 
   String? get _parsedName {
@@ -189,7 +189,7 @@ class _ThemedTimePickerState extends State<ThemedTimePicker> {
   }
 
   String t(String key, [Map<String, dynamic> args = const {}]) {
-    String result = LayrzAppLocalizations.of(context)?.t(key, args) ?? widget.translations[key] ?? key;
+    String result = LayrzAppLocalizations.maybeOf(context)?.t(key, args) ?? widget.translations[key] ?? key;
 
     if (widget.overridesLayrzTranslations) {
       result = widget.translations[key] ?? key;