+++ lib/src/theme/theme.dart
@@ -5,9 +5,13 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:layrz_models/layrz_models.dart';
 import 'package:layrz_theme/src/helpers/helpers.dart';
+import 'package:http/http.dart' as http;
 
 part 'src/platform.dart';
 part 'src/light_theme.dart';
 part 'src/dark_theme.dart';
 part 'src/constants.dart';
+part 'src/font.dart';